% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{annotClone}
\alias{annotClone}
\title{Annotate the clones by driver mutations}
\usage{
annotClone(
  x,
  y,
  annot,
  angle = 0,
  col = "black",
  pos = 4,
  cex = 0.5,
  offset = 0.5
)
}
\arguments{
\item{x}{graphical x position of the clone origin}

\item{y}{graphical y position of the clone origin}

\item{annot}{annotation/driver mutations}

\item{angle}{clone anotation angle}

\item{col}{clone annotation text colour}

\item{pos}{clone annotation (1=below, 2=left, 3=above, 4=right)}

\item{cex}{clone annotation text size}

\item{offset}{clone annotation text offset from start positoin}
}
\description{
Annotate the clones by driver mutations
}
