% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{checkCol}
\alias{checkCol}
\title{Check that the number of colors provided matches  up with the number of clones}
\usage{
checkCol(fish)
}
\arguments{
\item{fish}{A fish object}
}
\value{
no return value - stops execution if the numbers don't match up
}
\description{
Check that the number of colors provided matches  up with the number of clones
}
