% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{createBackgroundImage}
\alias{createBackgroundImage}
\title{Create the gradient background image for the plot}
\usage{
createBackgroundImage(col = NULL)
}
\arguments{
\item{col}{A vector of three colors to use for the gradient}
}
\value{
returns the location of the temporary png file that will get embedded into the eventual output
}
\description{
Create the gradient background image for the plot
}

