% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{drawClustPolygon}
\alias{drawClustPolygon}
\title{Draw a single cluster using polygons}
\usage{
drawClustPolygon(
  xpos,
  ytop,
  ybtm,
  color,
  nest.level,
  pad.left = 0,
  border = 1,
  col.border = NULL,
  ramp.angle = 0.5,
  annot = "",
  annot.angle,
  annot.col,
  annot.cex,
  annot.pos,
  annot.offset
)
}
\arguments{
\item{xpos}{A vector of x values for control points}

\item{ytop}{A vector of y values for control points on the top}

\item{ybtm}{A vector of y values for control points on the bottom}

\item{color}{A color value for this polygon}

\item{nest.level}{An integer describing how deeply this is nested}

\item{pad.left}{A numeric amount of extra padding to add to the left side of the shape}

\item{border}{A numeric width for the border line around this polygon}

\item{ramp.angle}{A numeric value between 0 and 1 that indicates how steeply the polygon should expand from it's origin to the first measured point}
}
\value{
No return value, outputs on graphics device
}
\description{
Draw a single cluster using polygons
}
\examples{
\dontrun{
drawClustPolygon(xpos=c(0,30,75,150), ytop=c(100,51,51,99), ybtm=c(0,49,49,1), color="red", nest.level=1)
}

}
