% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{fixDisappearingClones}
\alias{fixDisappearingClones}
\title{A clone that has a nonzero value at one time point cannot completely disappear in a second, and then reappear in a third.
That clone must have really been there all along.  This function will "fix" any such instances by replacing the in-between
zero values with a very small value.}
\usage{
fixDisappearingClones(frac.table, nest.level)
}
\arguments{
\item{frac.table}{A numeric matrix containing tumor fraction estimates for all clones at all timepoints}

\item{nest.level}{An integer vector specifying how deeply a given clone is nested in the overall hierarchy}
}
\value{
The matrix with appropriate zeros converted to appropriate small values
}
\description{
A clone that has a nonzero value at one time point cannot completely disappear in a second, and then reappear in a third.
That clone must have really been there all along.  This function will "fix" any such instances by replacing the in-between
zero values with a very small value.
}
\examples{
\dontrun{

frac.table = fixDisappearingClones(frac.table, nest.level)
}

}
