% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{getAllNestLevels}
\alias{getAllNestLevels}
\title{Given the a list representing the parents of each clone, return a vector specifying how deeply each clone is nested}
\usage{
getAllNestLevels(parents)
}
\arguments{
\item{parents}{An integer vector specifying parental relationships between clones}
}
\value{
An integer vector representing how deeply each subclone is nested
}
\description{
Given the a list representing the parents of each clone, return a vector specifying how deeply each clone is nested
}
\seealso{
getNestLevel
}
