% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{getInnerSpace}
\alias{getInnerSpace}
\title{Get the percentage of a clone that is only that clone and not occupied by subclones}
\usage{
getInnerSpace(clone, fish)
}
\arguments{
\item{clone}{The number of the clone to check}

\item{fish}{A fish object}
}
\value{
A number giving the percentage of this clone that is uniquely this clone
}
\description{
Get the percentage of a clone that is only that clone and not occupied by subclones
}

