% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{getNestLevel}
\alias{getNestLevel}
\title{Given the a list representing the parents of each clone, and the number specifying which clone to test, returns how deeply it is nested}
\usage{
getNestLevel(parents, x)
}
\arguments{
\item{parents}{An integer vector specifying parental relationships between clones}

\item{x}{The integer specifying which subclone to calculate nest level for}
}
\value{
An integer representing how deeply this subclone is nested
}
\description{
Given the a list representing the parents of each clone, and the number specifying which clone to test, returns how deeply it is nested
}
\seealso{
getAllNestLevels
}
