% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{getOuterSpace}
\alias{getOuterSpace}
\title{Get the amount of non-tumor space outside of all clones}
\usage{
getOuterSpace(fish)
}
\arguments{
\item{fish}{A fish object}
}
\value{
A numeric vector representing non-tumor space at each timepoint
}
\description{
Get the amount of non-tumor space outside of all clones
}

