% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{layoutClones}
\alias{layoutClones}
\title{Generate key points of the layout that will be used for plotting}
\usage{
layoutClones(fish, separateIndependentClones = FALSE)
}
\arguments{
\item{fish}{A fish object with appropriate slots filled (frac.table, parents, nest.level)}

\item{separateIndependentClones}{Boolean - Should independently-arising clones (with parent 0) be separated by blank space in the plot?}
}
\value{
A fish object with layout slots filled in
}
\description{
Generate key points of the layout that will be used for plotting
}
\examples{
\dontrun{
layoutClones(fish.object)

layoutClones(fish.object, separateIndependentClones=TRUE)
}

}

