% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{numChildren}
\alias{numChildren}
\title{get the number of non-zero children at a particular timepoint}
\usage{
numChildren(fish, clone, timepoint)
}
\arguments{
\item{fish}{A fish object}

\item{clone}{An integer representing the clone number to check}

\item{timepoint}{The timepoint at which to check}
}
\value{
the number of children with non-zero fractions
}
\description{
get the number of non-zero children at a particular timepoint
}

