% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_plotting_functions.R
\name{plot_OM_Ages}
\alias{plot_OM_Ages}
\title{Age structure of a population.}
\usage{
plot_OM_Ages(output2, region, skip)
}
\arguments{
\item{output2}{The output array from a model run.}

\item{region}{Which region should plotted?}

\item{skip}{Skip years if needed. = 1 means don't skip any years.}
}
\value{
Returns a histogram of age structure for every year.
}
\description{
Plot histogram of age frequencies for a particular region.
}
\details{
This function plots age frequency data from the \code{output2} array. \code{output2} should be of dimention [n_iters, ages, years, regions]
}
