% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_plotting_functions.R
\name{plot_OM_SSB}
\alias{plot_OM_SSB}
\title{SSB Time Series}
\usage{
plot_OM_SSB(output, item = "OM_ssb_R1", mean = TRUE, ...)
}
\arguments{
\item{output}{The output matrix from a model run.}

\item{item}{= OM_ssb_R1 as default but can be any known column name type.}

\item{mean}{= Logical. Should the mean be plotted?}
}
\value{
Returns a time series plot of item.
}
\description{
Plot SSB from output from Planetfish2 operating model or assessment model.
}
\details{
This function plots data from the \code{output} matrix. This can be either OM or AM output as SSB or REC etc.
The intention is to be able to plot an item by passing the unique column name identifier to \code{plot()}.
}
