% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{validateInputs}
\alias{validateInputs}
\title{Validate some key assumptions about the fish object's data}
\usage{
validateInputs(frac.table, parents, nest.level, clone.labels, clone.annots)
}
\arguments{
\item{frac.table}{A numeric matrix containing tumor fraction estimates for all clones at all timepoints}

\item{parents}{An integer vector specifying parental relationships between clones}

\item{nest.level}{An integer vector specifying how deeply a given clone is nested in the overall hierarchy}

\item{clone.labels}{An integer vector specifying how deeply a given clone is nested in the overall hierarchy}

\item{clone.annots}{A character vector of annotations (mutation) to label to each clone in the plot}
}
\value{
No return value - stops execution with an error if invalid inputs are detected
}
\description{
Validate some key assumptions about the fish object's data
}
