% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{createFishObject}
\alias{createFishObject}
\title{Create a fish object after doing some input validation and data munging}
\usage{
createFishObject(frac.table, parents, timepoints = NULL, col = NULL,
  clone.labels = NULL)
}
\arguments{
\item{frac.table}{A numeric matrix containing tumor fraction estimates for all clones at all timepoints}

\item{parents}{An integer vector specifying parental relationships between clones}

\item{timepoints}{An numeric vector specifying the timepoints for each column of the matrix}

\item{col}{A vector of colors to use when plotting each clone}
}
\value{
A fish object with the relevant slots filled
}
\description{
Create a fish object after doing some input validation and data munging
}
\examples{
timepoints=c(0,30,75,150)
frac.table = matrix(
    c(100, 45, 00, 00,
       02, 00, 00, 00,
       02, 00, 02, 01,
       98, 00, 95, 40),
    ncol=length(timepoints))
parents = c(0,1,1,3)
fish = createFishObject(frac.table,parents,timepoints=timepoints)

}

