% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishPlot.R
\name{fishPlot}
\alias{fishPlot}
\title{FishPlot}
\usage{
fishPlot(
  FishData,
  ladj = 0.001,
  jadj = 0.2,
  nadj = 0.01,
  Species = "Redfish",
  lab = NULL,
  yl = c(41, 47.85),
  xl = c(-68, -56.6),
  lscale = 30,
  type = "FishPlot"
)
}
\arguments{
\item{FishData}{The fish data}

\item{ladj}{The length adjustment}

\item{jadj}{The jitter adjustment}

\item{nadj}{The number adjustment}

\item{Species}{The Species Name}

\item{yl}{The latitudes}

\item{xl}{The longitudes}

\item{lscale}{The length scale}

\item{type}{The type of plot "mean.length" or "FishPlot"TM}

\item{yr}{The year}
}
\value{
A Fish Plot
}
\description{
FishPlot
}
