% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishplot.R
\docType{package}
\name{fishplot}
\alias{fishplot}
\alias{fishplot-package}
\title{fishplot: A package for plotting output from Planetfish2 and CASAL.}
\description{
The fishplot package contains various functions for plotting the output from OM or AM
models. This can be useful for diagnosing problems with the models or analyzing results.
The goal is to eventually add commonly used plotting functions for plots that will
be included in a manuscript.
}
\section{fishplot functions}{

plot_OM_SSB
plot_OM_ages
}

