% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{fishPlot}
\alias{fishPlot}
\title{Given a fish object containing layout information, draw the fish plot}
\usage{
fishPlot(fish, shape = "polygon", vlines = NULL, col.vline = "#FFFFFF99",
  vlab = NULL, border = 0.5, col.border = "#777777", pad.left = 0.2,
  ramp.angle = 0.5, title = NULL, title.btm = NULL, cex.title = NULL,
  cex.vlab = 0.7)
}
\arguments{
\item{fish}{A fish object that contains layout information}

\item{shape}{The type of shape to construct the plot out of. The "spline" and "polygon" methods work well. "bezier" is more hit or miss}

\item{vlines}{A vector of x positions at which to draw vertical lines}

\item{col.vline}{A color value for the vertical lines}

\item{vlab}{A character vector containing labels for each of the vertical lines}

\item{border}{A numeric width for the border line around this polygon}

\item{col.border}{A color for the border line}

\item{pad.left}{The amount of "ramp-up" to the left of the first timepoint. Given as a fraction of the total plot width. Default 0.2}

\item{ramp.angle}{A numeric value between 0 and 1 that indicates how steeply the shape should expand from it's leftmost origin to the first measured point. Only used when shape="polygon".}

\item{title}{A string for the title above the plot}

\item{title.btm}{A string for the title at the bottom left, internal to the plot}

\item{cex.title}{A numeric value for scaling the title size}

\item{cex.vlab}{A numeric value for scaling the top label size default is 0.7}
}
\value{
No return value, outputs on graphics device
}
\description{
Given a fish object containing layout information, draw the fish plot
}
\examples{
\dontrun{
fishPlot(fish,shape="polygon",title.btm="633734",
           vlines=c(0,150), vlab=c("day 0","day 150"), cex.title=0.5)
}
}

